convert stories page to tsx
